+++ Source/TeleCore/Static/StaticData.cs
@@ -106,13 +106,13 @@ public static class StaticData
     #region Def ID
 
     /// <summary>
-    ///     Returns the relative Def object assignable to the provided ID.
+    ///    Returns the unique Def instance of the provided ID.
     /// </summary>
     /// <param name="id">The ID of the Def.</param>
     /// <typeparam name="TDef">The Def type to search through.</typeparam>
     /// <returns>A unique Def instance as identified by the id.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static TDef ToDef<TDef>(this ushort id)
+    public static TDef ToDef<TDef>(this int id)
         where TDef : Def
     {
         return DefIDStack.ToDef<TDef>(id);
@@ -127,14 +127,14 @@ public static class StaticData
     }
 
     /// <summary>
-    ///     Turns a Def to its relative ID.
-    ///     <para>Can be used to handle Defs in a more lightweight way without assigning references to objects.</para>
+    ///     Returns the unique ID of the provided Def instance.
+    ///     <para>Can be used to handle Defs in a more lightweight way without assigning references to their instance.</para>
     /// </summary>
     /// <param name="def">The Def instance.</param>
     /// <typeparam name="TDef">The Def type to assign the ID from.</typeparam>
     /// <returns>A unique ID for the Def instance of the given Def type.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static ushort ToID<TDef>(this TDef def)
+    public static int ToID<TDef>(this TDef def)
         where TDef : Def
     {
         return DefIDStack.ToID(def); // def.index; //DefIDStack<TDef>.GetID(def);
